Make validity and completeness fields optional
Based on the quote from the TRT: 'The new test condition fields, “validity” and “completeness” have been added. For the 2014-2015
school year theyse values will be calculated if not present. In future years, the fields will be required.
Details about the calculation rules are in the “Test Condition Codes” document on SmarterApp.org.'